Move Traversal State to mtl and add docs
