+++ modules/optics/arrow-optics/build.gradle
@@ -7,6 +7,7 @@ dependencies {
     testRuntime("org.junit.vintage:junit-vintage-engine:$jUnitVintageVersion")
     testCompile "io.kotlintest:kotlintest-runner-junit4:$kotlinTestVersion"
     testCompile project(':arrow-test')
+    testCompile project(':arrow-optics-mtl')
 
     compile project(':arrow-annotations')
     kapt project(':arrow-meta')
